Assignment 2/assignment-2-da25m018-da25m019/requirements.txt
fastapi==0.129.0
uvicorn==0.41.0
requests==2.32.5
python-dotenv==1.2.1
gTTS==2.5.4
spacy==3.8.11